console: perf improvements
Transient index/vertex buffers moved to a C implementation that does not
suffer from boxing of values from Haxe.

Improved BitmapData->Texture mapping by keeping a pointer to the Texture
as a field in BitmapData for free lookup, as well as a pointer
associated with a WeakRef for automatic cleanup when BitmapData's are
GCed.